tests/test_transcriber.py
""" Test cases for the Transcriber class """
import unittest
from unittest.mock import MagicMock
from google.cloud.speech import RecognizeResponse
from module.transcriber import TranscriberTranslator

class TestTranscriberTranslator(unittest.TestCase):
    """ Test cases for the TranscriberTranslator class """

    def setUp(self):
        # Create an instance of TranscriberTranslator
        self.transcriber_translator = TranscriberTranslator()

    def test_transcribe_audio(self):
        """ Test transcribing audio data """
        audio_data = b'audio_data'
        expected_result = RecognizeResponse()
        self.transcriber_translator.speech_client.recognize = MagicMock(return_value=expected_result)
        result = self.transcriber_translator.transcribe_audio(audio_data)
        self.assertEqual(result, '')

    def test_translate_text(self):
        """ Test translating text """
        text = "Hello"
        expected_result = "Translated Text"
        self.transcriber_translator.translate_client.translate = MagicMock(return_value={"translatedText": expected_result})
        result = self.transcriber_translator.translate_text(text)
        self.assertEqual(result, expected_result)

    def test_transcribe_and_translate(self):
        """ Test transcribing and translating audio data """
        audio_data = b'audio_data'
        expected_result = "Translated Text"
        self.transcriber_translator.transcribe_audio = MagicMock(return_value="Transcribed Text")
        self.transcriber_translator.translate_text = MagicMock(return_value=expected_result)
        result = self.transcriber_translator.transcribe_and_translate(audio_data)
        self.assertEqual(result, expected_result)

if __name__ == '__main__':
    unittest.main()