requirements.txt
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-speech==2.26.1
google-cloud-translate==3.15.4
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
idna==3.7
logger==1.4
numpy==2.0.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
requests==2.32.3
rsa==4.9
scipy==1.14.0
sounddevice==0.4.7
urllib3==2.2.2
